+++ RPi-Triangle-Markers/robot.hh
@@ -17,16 +17,16 @@ class Robot
 
     public:
 
-        const float R=3.35;//cm
-		const float L=12.4;//cm distance between wheels
+        const double R=3.35;//cm
+        const double L=12.4;//cm distance between wheels
         static const int  WHEEL_RESOLUTION=20;
         static const int NUMWHEELS=2;
-        float A[2][2]={{L/(2*R),1/R}, {-L/(2*R),1/R}};//inverse matrix for compute the angular velocities of every wheel
-        void angularWheelSpeed(float (&w_wheel)[NUMWHEELS],float velocity_robot[2]);
+        double A[2][2]={{L/(2*R),1/R}, {-L/(2*R),1/R}};//inverse matrix for compute the angular velocities of every wheel
+        void angularWheelSpeed(double (&w_wheel)[NUMWHEELS],double velocity_robot[2]);
         void SetupRobotData(int,string,string);
         void SetupConection(int& ,string& ,string&);
         void linearVelocity();
-	    void angularVelocity();
+        void angularVelocity();
         
         void IMU();
